+++ solidity/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hyperlane-xyz/core
 
+## 9.0.4
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@16.1.1
+
+## 9.0.3
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@16.1.0
+
 ## 9.0.2
 
 ### Patch Changes