+++ typescript/eslint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @hyperlane-xyz/eslint-config
 
+## 16.1.1
+
+## 16.1.0
+
 ## 16.0.0
 
 ### Patch Changes